M7.7: four ratified C-ABI freeze-shape decisions (pre-certification)
The freeze-shape half of the M7.7 MINOR/NIT batch — the choices that are
irreversible once the ABI freezes at certification. All four ratified 2026-09-08.

1. DoodlePosition gains a `reserved` tail ([u32; 1]). It was the lone by-value
   struct without one, and it is embedded by value in DoodleFrame, so a wrong
   freeze would cascade. The tail is expected to stay zero forever (a position is
   deliberately just (module, byte-span), E§8.1) — recorded as commentary, not an
   exemption. Freeze convention 2 now mandates a reserved tail on EVERY by-value
   struct (no exceptions) and pins the mechanism: producer writes zero, consumer
   ignores, repurposable only at a minor ABI-version bump.

2. List-accessor width unified to u32. The callback surface
   (doodle_call_list_length/_get) used usize/uintptr_t for the same operation the
   inspection surface does in u32; u32 is the engine's real width (heap indices are
   u32, machine-design ground rule 2), so uintptr_t was the fiction. Convention 6
   now states: element counts/indices cross as fixed-width u32; raw byte-buffer
   sizes stay size_t (the copy-out cap/out_len — a memory quantity).

3. doodle_retain added (E§4.2's refcounted model; the surface had only release).
   Signature `doodle_retain(instance, handle) -> handle` (chainable, CFRetain/AddRef
   idiom); DOODLE_NULL_HANDLE on a stale/forged handle or a caught panic (a new
   guard::catch_or firewall). doodle_release's "as many times as obtained" doc is
   now accurate (a reference is obtained by minting OR retain). Test covers
   retain/release/release balance, use-after-final-release, and that
   retain-after-release does not resurrect (→ null).

4. DOODLE_GC_STRESS env read moved behind an off-by-default `gc-stress` cargo
   feature (doodle-capi + conformance-runner): the shipped library must not silently
   vary with an ambient variable. Only the env read is gated;
   Instance::enable_gc_stress() stays ungated. The gate
   (scripts/gc-stress-conformance.sh) and the Miri job now build
   `--release --features gc-stress` explicitly (so it can't degrade to debug-only);
   tests/gc_stress.rs is gated on the feature and runs under Miri with it.

+++ crates/doodle-capi/src/call_value.rs
@@ -254,7 +254,9 @@ pub unsafe extern "C" fn doodle_call_make_foreign(
 }
 
 /// A fresh **host-owned** handle to the element at `index` of a list (E§4.6). `ErrWrongKind` if
-/// not a list; `ErrIndexOutOfBounds` past the end.
+/// not a list; `ErrIndexOutOfBounds` past the end. `index` is a `u32`: element indices cross the
+/// ABI as fixed-width `u32` (the engine's heap is `u32`-indexed, machine-design ground rule 2),
+/// matching `doodle_list_get`.
 ///
 /// # Safety
 /// `ctx` live; `out` writable.
@@ -262,12 +264,12 @@ pub unsafe extern "C" fn doodle_call_make_foreign(
 pub unsafe extern "C" fn doodle_call_list_get(
     ctx: *mut DoodleCallCtx,
     list: DoodleHandle,
-    index: usize,
+    index: u32,
     out: *mut DoodleHandle,
 ) -> DoodleStatus {
     make_fallible(
         ctx,
-        |engine| engine.list_get(Handle::from_bits(list), index),
+        |engine| engine.list_get(Handle::from_bits(list), index as usize),
         out,
     )
 }
